fix: resolve lint errors and sandbox transport selection
- Add `_ =` to unchecked CloseWithError calls in hysteria2 dialer
- Add `_ =` to unchecked json.Encode calls in httpbin test helper
- Add ReadHeaderTimeout to httpbin http.Server (gosec G114)
- Suppress unchecked recover() return in server event bus
- Remove unnecessary uint64 conversions in webrtc stats
- Remove unused recvNonce field from P2PConn
- Remove unused trickleGatherer field from Manager
- Switch sandbox clients from "auto" to "reality" transport to avoid
  H3/UDP timeouts in Docker networks

+++ transport/webrtc/stats.go
@@ -59,8 +59,8 @@ func (sc *statsCollector) collect() {
 		case webrtc.ICECandidatePairStats:
 			if v.Nominated {
 				stats.RTT = time.Duration(v.CurrentRoundTripTime * float64(time.Second))
-				stats.BytesSent = uint64(v.BytesSent)
-				stats.BytesRecv = uint64(v.BytesReceived)
+				stats.BytesSent = v.BytesSent
+				stats.BytesRecv = v.BytesReceived
 				stats.PacketsSent = uint64(v.PacketsSent)
 				stats.PacketsRecv = uint64(v.PacketsReceived)
 			}
